Disable nginx_brotli
For some unknown reason its compilation is not working on Ubuntu 20.04
when running on Digital Ocean. In the Vagrant box it works just fine.

Since it is a compression module, which I judge not to be fundamental
for a test instance, it is ok to disable it just for our installation.
This is not intended to be merged into the upstream.